move tx bals
