+++ svm/src/lib.rs
@@ -8,6 +8,7 @@ pub mod nonce_info;
 pub mod program_loader;
 pub mod rollback_accounts;
 pub mod transaction_account_state_info;
+pub mod transaction_balances;
 pub mod transaction_commit_result;
 pub mod transaction_error_metrics;
 pub mod transaction_execution_result;
