+++ svm/src/transaction_balances.rs
@@ -1,0 +1,17 @@
+pub type TransactionBalances = Vec<Vec<u64>>;
+
+#[derive(Clone, Debug)]
+pub struct TransactionBalancesSet {
+    pub pre_balances: TransactionBalances,
+    pub post_balances: TransactionBalances,
+}
+
+impl TransactionBalancesSet {
+    pub fn new(pre_balances: TransactionBalances, post_balances: TransactionBalances) -> Self {
+        assert_eq!(pre_balances.len(), post_balances.len());
+        Self {
+            pre_balances,
+            post_balances,
+        }
+    }
+}
